+++ CHANGES.rst
@@ -6,6 +6,7 @@
 - Remove bundled ``configobj`` package in favor of the ``configobj`` package
   bundled into ``astropy``. [#122]
 - Fix ``datetime.utcnow()`` DeprecationWarning [#134]
+- Fix bug in handling of ``pathlib.Path`` objects as ``Step`` inputs [#143]
 
 0.5.1 (2023-10-02)
 ==================